+++ CHANGELOG.md
@@ -175,6 +175,7 @@ In order to generate Open Data exports you should add this to your crontab or re
 - **decidim-budgets**: Use bigint instead of int for projects [\#4628](https://github.com/decidim/decidim/pull/4628)
 - **decidim-comments**: Fix GraphQL comments schema [\#4638](https://github.com/decidim/decidim/pull/4638)
 - **decidim-core**: Better handling setting process step position [\#4638](https://github.com/decidim/decidim/pull/4638)
+- **decidim-core**: Comment authors won't be notified about their own comments [\#4643](https://github.com/decidim/decidim/pull/4643)
 - **decidim-core**: Don't crash on registration [\#4637](https://github.com/decidim/decidim/pull/4637)
 - **decidim-debates**: Don't crash on settings change [\#4642](https://github.com/decidim/decidim/pull/4642)
 - **decidim-proposals**: Don't crash on settings change [\#4642](https://github.com/decidim/decidim/pull/4642)